// charting_library/broker-api.d.ts
export declare const enum ConnectionStatus {
	Connected = 1,
	Connecting = 2,
	Disconnected = 3,
	Error = 4
}
export declare const enum ConnectionType {
	Demo = 1,
	Real = 0
}
export declare const enum NotificationType {
	Error = 0,
	Success = 1
}
export declare const enum OrderStatus {
	Canceled = 1,
	Filled = 2,
	Inactive = 3,
	Placing = 4,
	Rejected = 5,
	Working = 6
}
export declare const enum OrderStatusFilter {
	All = 0,
	Canceled = 1,
	Filled = 2,
	Inactive = 3,
	Rejected = 5,
	Working = 6
}
export declare const enum OrderTicketFocusControl {
	LimitPrice = 1,
	StopPrice = 2,
	TakeProfit = 3,
	StopLoss = 4,
	Quantity = 5
}
export declare const enum OrderType {
	Limit = 1,
	Market = 2,
	Stop = 3,
	StopLimit = 4
}
export declare const enum ParentType {
	Order = 1,
	Position = 2,
	Trade = 3
}
export declare const enum PriceType {
	Limit = 1,
	Stop = 2
}
export declare const enum Side {
	Buy = 1,
	Sell = -1
}
export declare const enum StandardFormatterName {
	Date = "date",
	DateOrDateTime = "dateOrDateTime",
	Default = "default",
	Fixed = "fixed",
	VariablePrecision = "variablePrecision",
	FormatQuantity = "formatQuantity",
	FormatPrice = "formatPrice",
	FormatPriceForexSup = "formatPriceForexSup",
	IntegerSeparated = "integerSeparated",
	LocalDate = "localDate",
	LocalDateOrDateTime = "localDateOrDateTime",
	Percentage = "percentage",
	Pips = "pips",
	Profit = "profit",
	ProfitInInstrumentCurrency = "profitInInstrumentCurrency",
	Side = "side",
	PositionSide = "positionSide",
	Status = "status",
	Symbol = "symbol",
	Text = "text",
	Type = "type",
	MarginPercent = "marginPercent",
	Empty = "empty"
}
// ...
export interface AccountManagerInfo {
	/** Name of the broker */
	accountTitle: string;
	/** Custom fields which will always be displayed above the pages. */
	summary: AccountManagerSummaryField[];
	/**
	 * Optional array to define custom formatters.
	 * Each description is an object with the following fields:
	 *
	 * 1. `name`: FormatterName
	 *     - Unique name of a formatter.
	 *
	 * 1. `formatText`: [TableFormatTextFunction](#tableformattextfunction)
	 *    -Function that is used for formatting of a cell value to `string`. Required because used to generate exported data.
	 *
	 * 1. `formatElement`: [CustomTableFormatElementFunction](#customtableformatelementfunction) | undefined
	 *     - Optional function that is used for formatting of a cell value to `string` or `HTMLElement`.
	 *
	 * If the `formatElement` function is provided, then only it will be used to format the displayed values, otherwise
	 * `formatText` will be used. If you need to only display `string` values it is better to use only `formatText` for performance reasons.
	 *
	 * **Example**
	 * ```ts
	 * {
	 *     name: 'closeButton' as FormatterName, // Typecast to FormatterName. Use constant in real code
	 *     formatText: () => '', // Returns an empty string because we don't need to display this in the exported data
	 *     formatElement: ({ values: [id] }: TableFormatterInputs<[id: string]>) => {
	 *         const button = document.createElement('button');
	 *
	 *         button.innerText = 'Close';
	 *
	 *         button.addEventListener('click', () => {
	 *             event.stopPropagation();
	 *
	 *             closePosition(id);
	 *         });
	 *
	 *         return button;
	 *     },
	 * }
	 * ```
	 */
	customFormatters?: CustomTableElementFormatter[];
	/** Columns description that you want to be displayed on the Orders page.
	 * You can display any field of an {@link Order}
	 * or add your own fields to an order object and display them.
	 */
	orderColumns: OrderTableColumn[];
	/** Optional sorting of the orders table. */
	orderColumnsSorting?: SortingParameters;
	/** History page will be displayed if it exists. All orders from previous sessions will be shown in the History. */
	historyColumns?: AccountManagerColumn[];
	/** Optional sorting of the history table. */
	historyColumnsSorting?: SortingParameters;
	/**
	 * You can display any field of a {@link Position}
	 * or add your own fields to a position object and display them.
	 */
	positionColumns?: AccountManagerColumn[];
	/**
	 * You can display any field of a {@link Trade}
	 * or add your own fields to a trade object and display them.
	 */
	tradeColumns?: AccountManagerColumn[];
	/** You can add new tabs in the Account Manager by using `pages`. Each tab is a set of tables. */
	pages: AccountManagerPage[];
	/** Optional list of statuses to be used in the orders filter. Default list is used if it hasn't been set. */
	possibleOrderStatuses?: OrderStatus[];
	/** Margin used */
	marginUsed?: IWatchedValue<number>;
	/**
	 * Optional function to create a custom context menu.
	 * @param contextMenuEvent - MouseEvent or TouchEvent object passed by a browser
	 * @param activePageActions - array of `ActionMetaInfo` items for the current page
	 * @returns `Promise` that is resolved with an array of `ActionMetaInfo`
	 */
	contextMenuActions?(contextMenuEvent: MouseEvent | TouchEvent, activePageActions: ActionMetaInfo[]): Promise<ActionMetaInfo[]>;
}
/** A description of an additional Account Manager tab. */
export interface AccountManagerPage {
	/** Unique identifier of a page */
	id: string;
	/** Page title. It is the tab name. */
	title: string;
	/** It is possible to display one or more tables in this tab. */
	tables: AccountManagerTable[];
}
/** Custom field that will always be shown above the pages of the Account manager */
export interface AccountManagerSummaryField {
	/** Text to display for the summary field */
	text: string;
	/** A WatchedValue object that can be used to read the state of field. */
	wValue: IWatchedValueReadonly<any>;
	/**
	 * Name of the formatter to be used for data formatting. If `formatter` is not
	 * set the value is displayed as is. Formatter can be a default or a custom one.
	 */
	formatter?: StandardFormatterName;
	/** Optional parameter which can be set to display the field by default. */
	isDefault?: boolean;
}
// ...
export interface AccountMetainfo {
	/**
	 * Account ID
	 */
	id: AccountId;
	/**
	 * Account Name
	 */
	name: string;
	/**
	 * Currency
	 */
	currency?: string;
	/**
	 * Currency Sign (Prefix)
	 */
	currencySign?: string;
}
// ...
export interface ActionDescriptionWithCallback extends ActionDescription {
	/** Action to be executed when user clicks the menu item. */
	action: (a?: ActionDescription) => void;
}
export interface BaseInputFieldValidatorResult {
	/** Is the base input value valid */
	valid: boolean;
}
export interface BracketOrder extends BracketOrderBase, CustomFields {
}
export interface BracketOrderBase extends PlacedOrderBase {
	/** If order is a bracket then this should contain base order/position id. */
	parentId: string;
	/** Type of the bracket's parent */
	parentType: ParentType;
}
export interface Brackets {
	/** Stop loss */
	stopLoss?: number;
	/** Take Profit */
	takeProfit?: number;
	/** Trailing Stop Pips */
	trailingStopPips?: number;
}
// ...
export interface CustomFields {
	/** Custom field */
	[key: string]: any;
}
// ...
export interface CustomTableElementFormatter<T extends TableFormatterInputValues = TableFormatterInputValues> {
	/** Custom formatter name */
	name: FormatterName;
	/** Formatter to generate HTML element */
	formatElement?: CustomTableFormatElementFunction<T>;
	/** Formatter to generate text. Return an empty string if you don't need to display this */
	formatText: TableFormatTextFunction<T>;
}
// ...
export interface DefaultContextMenuActionsParams {
}
// ...
export interface Execution extends CustomFields {
	/** Symbol name */
	symbol: string;
	/** Execution price */
	price: number;
	/** Execution Quantity */
	qty: number;
	/** Execution Side */
	side: Side;
	/** Time (unix timestamp in milliseconds) */
	time: number;
	/** Commission amount for executed trade */
	commission?: number;
	/** Net amount for executed trade */
	netAmount?: number;
}
// ...
export interface IDelegate<TFunc extends Function> extends ISubscription<TFunc> {
	/** Fire (Evoke) */
	fire: TFunc;
}
// ...
export interface IObservableValue<T> extends IBoxedValue<T>, IObservable<T> {
}
// ...
export interface IWatchedValue<T> extends IWatchedValueReadonly<T>, IObservableValue<T> {
	/**
	 * Set value for the watched value
	 * @param  {T} value - value to set
	 * @param  {boolean} forceUpdate? - force an update
	 */
	setValue(value: T, forceUpdate?: boolean): void;
	/** @inheritDoc */
	subscribe(callback: WatchedValueCallback<T>, options?: WatchedValueSubscribeOptions): void;
	/** @inheritDoc */
	unsubscribe(callback?: WatchedValueCallback<T> | null): void;
}
// ...
export interface InstrumentInfo {
	/** Quantity field step and boundaries */
	qty: QuantityMetainfo;
	/** Value of 1 pip for the instrument in the account currency */
	pipValue: number;
	/** Size of 1 pip (e.g., 0.0001 for EURUSD) */
	pipSize: number;
	/** Minimal price change (e.g., 0.00001 for EURUSD). Used for price fields. */
	minTick: number;
	/** Lot size */
	lotSize?: number;
	/** Instrument type. `forex` enables negative pips. You can check that in the Order dialog. */
	type?: SymbolType;
	/** Units of quantity or amount. Displayed instead of the Units label in the Quantity/Amount field. */
	units?: string;
	/** Display name for the symbol */
	brokerSymbol?: string;
	/** A description to be displayed in the UI dialogs. */
	description: string;
	/** Number of decimal places of DOM asks/bids volume (optional, 0 by default). */
	domVolumePrecision?: number;
	/** Leverage */
	leverage?: string;
	/**
	 * The margin requirement for the instrument. A 3% margin rate should be represented as 0.03.
	 * @deprecated
	 */
	marginRate?: number;
	/** Minimal price change for limit price field of the Limit and Stop Limit order. If set it will override the `minTick` value. */
	limitPriceStep?: number;
	/** Minimal price change for stop price field of the Stop and Stop Limit order. If set it will override the `minTick` value. */
	stopPriceStep?: number;
	/** Array of strings with valid duration values. You can check that in the Order dialog. */
	allowedDurations?: string[];
	/**
	 * Dynamic minimum price movement.
	 * It is used if the instrument's minimum price movement changes depending on the price range.
	 *
	 * For example: `0.01 10 0.02 25 0.05`, where `minTick` is `0.01` for a price less than `10`, `minTick` is `0.02` for a price less than `25`, `minTick` is `0.05` for a price more and equal than `25`.
	 */
	variableMinTick?: string;
	/** Instrument currency that is displayed in the Order dialog */
	currency?: string;
	/** The first currency quoted in a currency pair. Used for crypto currencies only. */
	baseCurrency?: string;
	/** The second currency quoted in a currency pair. Used for crypto currencies only. */
	quoteCurrency?: string;
	/** The value represented by a full point of price movement in the contract currency. This value is used to calculate the Total Value (symbol currency) of the order. */
	bigPointValue?: number;
	/** The value represents how much price is multiplied in relation to base monetary unit. */
	priceMagnifier?: number;
}
// ...
export interface MenuSeparator extends ActionDescription {
	/** Is a menu separator */
	separator: boolean;
}
// ...
export interface PlaceOrderResult {
	/** Order id */
	orderId?: string;
}
export interface PlacedOrder extends PlacedOrderBase, CustomFields {
}
// ...
export interface Position extends PositionBase, CustomFields {
}
// ...
export interface PreOrder extends OrderTemplate {
	/** @inheritDoc */
	symbol: string;
	/** @inheritDoc */
	type: OrderType;
	/** @inheritDoc */
	side: Side;
	/** @inheritDoc */
	qty: number;
	/** Current Quotes */
	currentQuotes?: AskBid;
	/**
	 * It is set to `true`, if the order closes a position.
	 */
	isClose?: boolean;
}
// ...
export interface SortingParameters {
	/** `property` of the data object that will be used for sorting */
	property: string;
	/** Ascending sorting order (default `true`) - If it is `false`, then initial sorting will be in descending order */
	asc?: boolean;
}
// ...
export interface TradeContext {
	/** Symbol name */
	symbol: string;
	/** Symbol display name */
	displaySymbol: string;
	/** Price value */
	value: number | null;
	/** Formatted value */
	formattedValue: string;
	/** Previous value */
	last: number;
}
// ...
export type AccountId = Nominal<string, "AccountID">;
export type AccountManagerColumn = {
	[K in StandardFormatterName | FormatterName]: AccountManagerColumnBase<K>;
}[StandardFormatterName | FormatterName];
/**
 * Describes a single action to put it into a dropdown or a context menu.
 */
export type ActionMetaInfo = ActionDescriptionWithCallback | MenuSeparator;
// ...
export type Order = PlacedOrder | BracketOrder;
export type OrderTableColumn = AccountManagerColumn & {
	/**
	 * An optional numeric array of order statuses that is applied to order columns only. If it is available then the column will be displayed in the specified tabs of the status filter only.
	 *
	 * Here is the list of possible order statuses:
	 *
	 * 0 - All
	 * 1 - Canceled
	 * 2 - Filled
	 * 3 - Inactive
	 * 5 - Rejected,
	 * 6 - Working
	 */
	supportedStatusFilters?: OrderStatusFilter[];
};
// ...
export type WatchedValueCallback<T> = (value: T) => void;